solar_energy_monitor/requirements.txt
web3==5.29.0
solcx==0.8.7
python-dotenv==1.0.0
